Local changes commit before pulling from upstream
